+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # NoveltyProducer - Changelog
 
+## 0.0.2 (2018-08-27)
+
+* Added .travis.yml, tests/test_generator.py, test-requirements.txt and setup.py.
+* Major changes in Generator.py:
+    - Added several Exception classes.
+    - Added optional input seed for random number generator.
+    - Added `_check_input` to check given inputs.
+    - Modified `_get_noise` and `get_data` to work with optional times_ input.
+    - Modified `get_payload` and `_get_times` accordingly.
+    - Added `_plant_a_seed` to (re)plant seed into random number generator.
+* Fixed typo in NoveltyProducer.py.
+
 ## 0.0.1 (2018-08-22)
 
 * Added CHANGELOG.md.